Lab1/requirement.txt
asgiref==3.7.2
contourpy==1.1.1
cycler==0.11.0
Django==4.2.5
fonttools==4.42.1
h5py==3.9.0
image==1.5.33
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
